move kubernetes tests to the generator folder
